+++ src/main/java/com/produtos/apirest/service/UsuarioService.java
@@ -3,18 +3,21 @@ package com.produtos.apirest.service;
 import com.produtos.apirest.models.Usuario;
 import com.produtos.apirest.repository.UsuarioRepo;
 import com.produtos.apirest.service.excecoes.RegraNegocioRunTime;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UsuarioService {
 
-    @Autowired
-    private UsuarioRepo usuarioRepo;
+    private final UsuarioRepo usuarioRepo;
+
+    public UsuarioService(UsuarioRepo usuarioRepo){
+        this.usuarioRepo = usuarioRepo;
+    }
 
     private BCryptPasswordEncoder passwordEncoder(){
         return new BCryptPasswordEncoder();
@@ -32,13 +35,13 @@ public class UsuarioService {
     }
 
     public static void verificaId(Usuario usuario){
-        if (usuario == null || Long.valueOf(usuario.getUsuarioId()) == null){
+        if (usuario == null || usuario.getUsuarioId() <= 0){
             throw new RegraNegocioRunTime("Usuario deve possuir um identificador!");
         }
     }
 
     public static void verificaId(Long id){
-        if (Long.valueOf(id) == null){
+        if (id <= 0){
             throw new RegraNegocioRunTime("Usuario deve possuir um identificador!");
         }
     }
@@ -76,15 +79,20 @@ public class UsuarioService {
     @Transactional
     public Usuario removerComFeedback(Long id){
         verificaId(id);
-        Usuario usuarioFeedback = usuarioRepo.findById(id).get();
-        usuarioRepo.delete(usuarioFeedback);
-        return usuarioFeedback;
+        Optional<Usuario> usuariosEncontrados = usuarioRepo.findById(id);
+        if (usuariosEncontrados.isPresent()) {
+            Usuario usuarioFeedback = usuariosEncontrados.get();
+            usuarioRepo.delete(usuarioFeedback);
+            return usuarioFeedback;
+        }
+        return null;
     }
 
     @Transactional
     public Usuario buscarPorId(Long id){
         verificaId(id);
-        return usuarioRepo.findById(id).get();
+        Optional<Usuario> usuariosEncontrados = usuarioRepo.findById(id);
+        return usuariosEncontrados.orElse(null);
     }
 
     @Transactional
